% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{ambig}
\alias{ambig}
\title{ambig}
\usage{
ambig(sol)
}
\arguments{
\item{sol}{QCA solution}
}
\value{
returns FALSE if the solution is not ambiguous
if the solution is ambiguous, the function prints out
the length of the ambiguity
}
\description{
A function that detects if a QCA solution is ambiguous
}
\details{
subsets of consistent and inconsistent rows.
}
