% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_div.R
\name{diversity}
\alias{diversity}
\title{Diversity of cases belonging to the same partition of the pooled data}
\usage{
diversity(x, units, time, cond, out, n_cut, incl_cut, BE_cons, WI_cons)
}
\arguments{
\item{x}{Calibrated pooled dataset for partitioning and minimization}

\item{units}{Units defining the within-dimension of data (time series)}

\item{time}{Periods defining the between-dimension of data (cross sections)}

\item{cond}{Conditions used for the pooled analysis}

\item{out}{Outcome used for the pooled analysis}

\item{n_cut}{Frequency cut-off for designating truth table rows as observed}

\item{incl_cut}{Inclusion cut-off for designating truth table rows as
consistent}

\item{BE_cons}{Inclusion (or consistency) thresholds for cross sections. 
Must be specified as a numeric vector with length equaling the number of
cross sections. Numbers correspond to the order of the cross section ID
in the data (such as years in ascending order).}

\item{WI_cons}{Inclusion (or consistency) thresholds for time series. 
Must be specified as a numeric vector with length equaling the number of
time series. Numbers correspond to the order of the time series (unit) ID
in the data (such as countries in alphabetical order).}
}
\value{
A dataframe presenting the diversity of cases belonging to the
same partition.
}
\description{
\code{diversity} calculates the diversity of cases that belong to the same
partition of the clustered data (time series, cross section etc.).
Diversity is measured by the number of truth table rows that the cases
cover. It calculates diversity across all truth table rows and for the
subsets of consistent and inconsistent rows.
}
\examples{
data(schwarz2016)
Schwarz_diversity <- diversity(schwarz2016, units = "country", time = "year", 
cond = c("poltrans", "ecotrans", "reform", "conflict", "attention"), 
out = "enlarge", 1, 0.8)

}
