% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Grauvogel2014}
\alias{Grauvogel2014}
\title{Original data used by Grauvogel/von Soest (2014)}
\format{
A data frame with 120 rows and 10 variables:
\describe{
  \item{Code}{Sender-target ID}
  \item{Sender}{Country or institution imposing sanctions}
  \item{Target}{Country that is target of sanctions}
  \item{Timeframe}{Considered years for each country case}
  \item{Persistence}{Degree of regime persistence after the intervention}
  \item{Comprehensiveness}{Scope of the imposed sanctions - comprehensive 
  vs. targeted sanctions}
  \item{Linkage}{Economic and social, respectively communicative and 
  geographic ties}
  \item{Vulnerability}{Military and economic vulnerability of the state to 
  outside pressure}
  \item{Repression}{Degree of repression by the state}
  \item{Claims}{Variety and strength of claims to legitimacy}
}
}
\source{
Grauvogel (2014) <doi:10.1111/1475-6765.12065>
}
\usage{
Grauvogel2014
}
\description{
A dataset containing the calibrated set values for the article:
Grauvogel, Julia and Christian von Soest (2014):  
Claims to Legitimacy Count: Why Sanctions Fail to Instigate 
Democratisation in Authoritarian Regimes.  
European Journal of Political Research 53 (4): 635-653.
}
\keyword{datasets}
