% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Schwarz2016}
\alias{Schwarz2016}
\title{Original data used by Schwarz (2016)}
\format{
A data frame with 74 rows and 9 variables:
\describe{
  \item{Case.ID}{Country-year ID}
  \item{enlarge}{Progress in the EU accession process}
  \item{poltrans}{Democracy status of the country}
  \item{ecotrans}{Market economy status of the country}
  \item{reform}{State of reform policy}
  \item{conflict}{Mean conflict intensity in a country per year}
  \item{attention}{EU’s attention to the issue of enlargement}
  \item{year}{Year ID}
  \item{country}{Country ID}
}
}
\source{
Schwarz (2016) <doi:10.1080/07036337.2016.1203309>
}
\usage{
Schwarz2016
}
\description{
A dataset containing the calibrated set values for the article:
Schwarz, Oliver (2016): Two Steps Forward One Step Back: 
What Shapes the Process of EU Enlargement in South-Eastern Europe? 
Journal of European Integration 38 (7): 757-773.
}
\keyword{datasets}
