% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ambiguity.R
\name{select.fnc}
\alias{select.fnc}
\title{select.fnc}
\usage{
select.fnc(sol, elmnt, overall = FALSE)
}
\arguments{
\item{sol}{QCA solution}

\item{elmnt}{is specific part of the solution that}

\item{overall}{a logical argument that allows to move between different 
levels of ambiguous solution. When FALSE returns elements per each model
within an ambiguous solution. When TRUE returns overall values
is needed (as saved by QCA package)}
}
\value{
returns coverage and consistency information
for each part of the ambiguous solution
}
\description{
A function that selects element pertaining
to each part of ambiguous solutions
}
