% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Thiem2011}
\alias{Thiem2011}
\title{Original data used by Thiem (2011)}
\format{
A data frame with 165 rows and 10 variables:
\describe{
  \item{id}{Country-year ID}
  \item{year}{Time ID}
  \item{country}{Country ID}
  \item{memberfs}{Monadic count of membership in formal intergovernmental 
  agreements on armaments cooperation}
  \item{fedismfs}{Degree to which a country’s domestic constitutional setup 
  is federalist in character}
  \item{homogtyfs}{Bilateral interaction scores based on all UN and NATO 
  military missions conducted between 1996 and 2006}
  \item{powdifffs}{Score to measure a country's military power based on the 
  CINC score}
  \item{comptvnsfs}{Competitiveness of a country’s domestic armaments 
  industry}
  \item{pubsupfs}{Public support for cooperation in defence}
  \item{ecodpcefs}{Degree of economic dependence}
}
}
\source{
Thiem(2011) <doi:10.1017/S1755773910000251>
}
\usage{
Thiem2011
}
\description{
A dataset containing the calibrated set values for the article:
Thiem, Alrik (2011): 
Conditions of Intergovernmental Armaments Cooperation in Western Europe,
 1996-2006. European Political Science Review 3 (1): 1-33.
}
\keyword{datasets}
