% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_conf.R
\name{upset_conf}
\alias{upset_conf}
\title{UpSetR for configurations}
\usage{
upset_conf(df, nsets)
}
\arguments{
\item{df}{Dataframe created with \code{\link{partition_min}} or
\code{\link{partition_min_inter}}.}

\item{nsets}{Specifies number of sets to be plotted. Argument
imported from the \code{\link{upset}} function from \pkg{UpSetR}.}
}
\value{
A plot presenting the frequency of individual
terms and their cooccurrences across QCA solutions.
}
\description{
UpSetR for configurations
}
