% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset.R
\name{upset_plot}
\alias{upset_plot}
\title{Aggregation of partition-specific models

Modelds that have been derived for individual partitions can be aggregated
to find out whether they share common elements.}
\usage{
upset_plot(df, nsets)
}
\arguments{
\item{df}{Dataframe created with \code{\link{partition_min}} or
\code{\link{partition_min_inter}}.}

\item{nsets}{Number of terms to include in plot}
}
\description{
Aggregation of partition-specific models

Modelds that have been derived for individual partitions can be aggregated
to find out whether they share common elements.
}
