% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_min.R
\name{partition_min}
\alias{partition_min}
\title{Generation of conservative or parsimonious solution for individual
partitions}
\usage{
partition_min(
  x,
  units,
  time,
  cond,
  out,
  n_cut,
  incl_cut,
  solution,
  BE_cons,
  WI_cons
)
}
\arguments{
\item{x}{Calibrated pooled dataset that is partitioned and minimized for
deriving the pooled solution.}

\item{units}{Units defining the within-dimension of data (time series)}

\item{time}{Periods defining the between-dimension of data (cross sections)}

\item{cond}{Conditions used for minimization}

\item{out}{Outcome used for minimization}

\item{n_cut}{Frequency cut-off for designating truth table rows as observed
as opposed to designating them as remainders.}

\item{incl_cut}{Inclusion (a.k.a. consistency) cut-off for designating
truth table rows as consistent.}

\item{solution}{A character specifying the type of solution that should
be derived. \code{C} produces the conservative (or complex) solution,
\code{P} for the parsimonious solution. See \code{\link{partition_min_inter}}
for a separate function for the intermediate solution.}

\item{BE_cons}{Inclusion thresholds for creating an individual truth table
for each cross section.
They must be specified as a numeric vector. Its length should be
equal the number of cross sections. The order of thresholds corresponds
to the order of the cross sections in the data defined by the cross-section
ID in the dataset (such as years in ascending order).}

\item{WI_cons}{Inclusion thresholds for creating an individual truth table
for each time series.
They must be specified as a numeric vector. Its length should be
equal the number of time series. The order of thresholds corresponds
to the order of the of the time-series (unit) ID
in the dataset (such as countries in alphabetical order).}
}
\value{
A dataframe summarizing the partition-specific and pooled solutions
with the following columns:
\itemize{
\item type: The type of the partition. \code{pooled} are rows with information
on the pooled data; \code{between} is for cross-section partitions;
\code{within} is for time-series partitions
\item consistency
}
}
\description{
\code{partition_min} decomposes clustered data into individual
partitions. For panel data, for example, these can be cross sections,
time series or both. The function derives an individual solution for
each partition and the pooled data to assess the robustness of the
solutions in a comparative perspective.
}
\examples{
data(Thiem2011)


}
