% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_configurations.R
\name{upset_conf}
\alias{upset_conf}
\title{Aggregation of individual configurations  over partition-specific models}
\usage{
upset_conf(df, nsets)
}
\arguments{
\item{df}{Dataframe created with \code{\link{partition_min}} or
\code{\link{partition_min_inter}}.}

\item{nsets}{Number of sets to include in plot (default is 5).}
}
\value{
An UpSet plot produced with \code{\link[UpSetR]{upset}}.
}
\description{
Models that have been derived for individual partitions are first
decomposed into sufficient terms, that is single sufficient conditions or
configurations. The individual terms are aggregated using UpSet plots to
determine how frequent they are individually and in combination.
}
