% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wop.R
\name{wop}
\alias{wop}
\title{Calculation of weight of partitions for pooled solution parameters}
\usage{
wop(
  dataset,
  units,
  time,
  cond,
  out,
  n_cut,
  incl_cut,
  solution,
  BE_cons,
  WI_cons,
  BE_ncut,
  WI_ncut
)
}
\arguments{
\item{units}{Units defining the within-dimension of data (time series)}

\item{time}{Periods defining the between-dimension of data (cross sections)}

\item{cond}{Conditions used for the pooled analysis}

\item{out}{Outcome used for the pooled analysis}

\item{n_cut}{Frequency cut-off for designating truth table rows as observed}

\item{incl_cut}{Inclusion cut-off for designating truth table rows as
consistent}

\item{solution}{A character specifying the type of solution that should
be derived. "C" produces the conservative (or complex) solution, "P" the
parsimonious solution. See \code{\link{wop_intermediate}} for the
intermediate solution.}

\item{BE_cons}{Inclusion (or consistency) thresholds for cross sections. 
Must be specified as a numeric vector with length equaling the number of
cross sections. Numbers correspond to the order of the cross section ID
in the data (such as years in ascending order).}

\item{WI_cons}{Inclusion (or consistency) thresholds for time series. 
Must be specified as a numeric vector with length equaling the number of
time series. Numbers correspond to the order of the time series (unit) ID
in the data (such as countries in alphabetical order).}

\item{BE_ncut}{For *cross sections*, the minimum number of members needed
for declaring a truth table row 
as relevant as opposed to designating it as a remainder.
Must be specified as a numeric vector. Its length should be
equal the number of cross sections. The order of thresholds corresponds
to the order of the cross sections in the data defined by the cross-section
ID in the dataset (such as years in ascending order).}

\item{WI_ncut}{For *time series*, the minimum number of members needed
for declaring a truth table row 
as relevant as opposed to designating it as a remainder. 
Must be specified as a numeric vector. Its length should be
equal the number of time series. The order of thresholds corresponds
to the order of the of the time-series (unit) ID
in the dataset (such as countries in alphabetical order).}

\item{x}{Calibrated pooled dataset for partitioning and minimization}
}
\value{
A dataframe with the weight of the partitions for pooled consistency
scores.
}
\description{
\code{wop} calculates the weight of partitions for each 
individual solution derived from the \code{\link{partition_min}}
function for pooled solution parameters. It derives a a score for the
numerator and denimonator for each solution of the clustered data.
}
\examples{

}
