% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wop.R
\name{wop}
\alias{wop}
\title{Weight of partitions for pooled solution parameters
for conservative or parsimonious solution}
\usage{
wop(dataset, units, time, cond, out, n_cut, incl_cut, solution, amb_selector)
}
\arguments{
\item{dataset}{Calibrated pooled dataset for partitioning and minimization
of pooled solution.}

\item{units}{Units that define the within-dimension of data (time series).}

\item{time}{Periods that define the between-dimension of data (cross sections).}

\item{cond}{Conditions used for the pooled analysis.}

\item{out}{Outcome used for the pooled analysis.}

\item{n_cut}{Frequency cut-off for designating truth table rows as observed
in the pooled analysis.}

\item{incl_cut}{Inclusion cut-off for designating truth table rows as
consistent in the pooled analysis.}

\item{solution}{A character specifying the type of solution that should
be derived. \code{C} produces the conservative (or complex) solution,
\code{P} the parsimonious solution. See \code{\link{wop_inter}} for deriving
intermediate solution.}

\item{amb_selector}{Numerical value for selecting a single model in the
presence of model ambiguity. Models are numbered according to their
order produced by \code{\link{minimize}} by the \code{QCA} package.}
}
\value{
A dataframe with information about the weight of the partitions
with the following columns:
\itemize{
\item \code{type}: The type of the partition. \code{pooled} are rows with
information on the pooled data; \code{between} is for
cross-section partitions; \code{within} is for time-series partitions.
\item \code{partition}: Specific dimension of the partition at hand. For
between-dimension, the unit identifiers are listed  (argument \code{units}).
For the within-dimension, the time identifier are listed (argument \code{time}).
The entry is \code{-} for the pooled data without partitions.
\item \code{solution}: The solution derived for the partition or the pooled data.
Absence of a condition is denoted by the \code{~} sign.
\item \code{model}: Running ID for models. In the presence of model ambiguity, each
model has its own row with its individual solution and parameters. The rest of
the information in the row is duplicated, for example by having two rows for
the within-partition 1996. The column \code{model} highlights the presence of
model ambiguity by numbering all models belonging to the same solution. For
example, if three consecutive rows are numbered 1, 2 and 3, then these rows
belong to the same solution and represent model ambiguity. If a 1 in a row
is followed by another 1, then there is no model ambiguity.
\item \code{num}: The contribution of the data to the numerator of the consistency
formula. For the pooled data, the value draws on all cases. For
partitions, the value represents the weight of the partition. The larger the
value, the larger the weight. The sum over all partition-specific weights is
equal to the value for the pooled data.
\item \code{denom}: The contribution of the data to the denominator of the
consistency formula. For the pooled data, the value draws on all cases.
For partitions, the value represents the weight of the partition.
The larger the value, the larger the weight. The sum over all
partition-specific weights is equal to the value for the pooled data.
}
}
\description{
\code{wop} calculates the contribution or weight of partitions
for the pooled solution parameters of consistency and coverage
for the conservative or parsimonious solution.
}
\examples{
data(Thiem2011)
Thiem_wop_pars <- wop(
  dataset = Thiem2011,
  units = "country", time = "year",
  cond = c("fedismfs", "homogtyfs", "powdifffs", "comptvnsfs", "pubsupfs", "ecodpcefs"),
  out = "memberfs",
  n_cut = 6, incl_cut = 0.8,
  solution = "P",
  amb_selector = 1)
Thiem_wop_pars
}
