% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wop.R
\name{wop}
\alias{wop}
\title{Calculation of weight of partitions for pooled solution parameters}
\usage{
wop(x, units, time, cond, out, n_cut, incl_cut, solution, BE_cons, WI_cons)
}
\arguments{
\item{x}{Calibrated pooled dataset for partitioning and minimization}

\item{units}{Units defining the within-dimension of data (time series)}

\item{time}{Periods defining the between-dimension of data (cross sections)}

\item{cond}{Conditions used for the pooled analysis}

\item{out}{Outcome used for the pooled analysis}

\item{n_cut}{Frequency cut-off for designating truth table rows as observed}

\item{incl_cut}{Inclusion cut-off for designating truth table rows as
consistent}

\item{solution}{A character specifying the type of solution that should
be derived. "C" produces the conservative (or complex) solution, "P" the
parsimonious solution. See \code{\link{wop_intermediate}} for the
intermediate solution.}

\item{BE_cons}{Inclusion (or consistency) thresholds for cross sections. 
Must be specified as a numeric vector with length equaling the number of
cross sections. Numbers correspond to the order of the cross section ID
in the data (such as years in ascending order).}

\item{WI_cons}{Inclusion (or consistency) thresholds for time series. 
Must be specified as a numeric vector with length equaling the number of
time series. Numbers correspond to the order of the time series (unit) ID
in the data (such as countries in alphabetical order).}
}
\value{
A dataframe with the weight of the partitions for pooled consistency
scores.
}
\description{
\code{wop} calculates the weight of partitions for each 
individual solution derived from the \code{\link{partition_min}}
function for pooled solution parameters. It derives a a score for the
numerator and denimonator for each solution of the clustered data.
}
\examples{

}
